Add files via upload
Job descriptions sorted and counted based on words in the text.  Additionally did similar word counts for industry, job tittle and region.   Industry may be interesting feature to display.   

The initial word text file for the description has "data" and "experience" as the top of the description list which I sorted out to improve visibility in the  plots/ charts.  Also similar plots for the industry most common was interesting.